Unique ID and server class
Created a server class which initializes the table and local file with
an unique identifier. If either of those already exist the server will
take on the unique identifier that already exists. Added an input
variable for all the functions which handle the table. Main() does not
work now as it is not changed to fit the Server class or updated
functions